feat(socket): rework inner networking (#3)
* feat(crate): migrate to `fastwebsockets` crate

* feat(util): add user agent helper fn

* chore(util): cleanup tcp proxy

* chore(structs): cleanup `handshake`

* feat(base): migrate to `hyper` request

* feat(socket): rework inner networking

* chore(socket): remove dev logging

+++ src/structs/handshake.rs
@@ -1,11 +1,11 @@
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, Clone, PartialEq, Eq, Hash, PartialOrd, Ord)]
 pub struct Handshake {
     pub sid: Option<String>,
     pub upgrades: Vec<String>,
     #[serde(rename = "pingTimeout")]
-    pub ping_timeout: u32,
+    pub ping_timeout: u64,
     #[serde(rename = "pingInterval")]
-    pub ping_interval: u32,
+    pub ping_interval: u64,
 }
